feeder/requirements.txt
Pillow==4.0.0
PyYAML==3.12
Unipath==1.1
coverage==4.3.4
gevent==1.2.1
greenlet==0.4.11
grequests==0.3.0
mock==2.0.0
mockredispy==2.9.3
nose==1.3.7
olefile==0.44
pbr==1.10.0
python-dateutil==2.6.0
redis==2.10.5
requests==2.12.5
six==1.10.0
